feat(cli): replace v0.1 add wholesale with v0.2 pipeline as default (#89)
Twelfth slice of Phase 3. Drops the transitional `--pipeline` flag and
makes the v0.2 SSOT generation pipeline the default behaviour of
`upskill add`. The shipped binary on `main` is still v0.1.x, so the
deletions here are local to the `v0.2-redesign` branch.

CLI:
- `add <source>` now runs `pipeline::install_with_lockfile` directly.
  No `--pipeline` flag, no per-agent `--claude` / `--copilot` / `--all`
  / `--copy`, no `--skill` filter (the pipeline installs every item the
  source contains, per ADR-0004).
- `remove`, `update`, `list`, `doctor` accept their final ADR-0004 arg
  shapes but currently print a "not yet implemented (lands in Phase Bx)"
  message and exit 1. They land over Phases B1-B3.
- `search` is unchanged.

Modules removed (v0.1 only): `src/install.rs`, `src/agent.rs`,
`src/lockfile.rs`, `src/ui.rs`. v0.1's `hash_skill_dir` migrates to
`pipeline::hash_item_dir` (private). The v0.1 lockfile deserialiser
shrinks to a private `V1Lockfile` struct inside `lockfile_v2.rs` so
in-place migration of v0.1 `.upskill-lock.json` still works.

Tests removed (v0.1 only): `cli_check`, `cli_dryrun`, `cli_gitlab`,
`cli_global`, `cli_list`, `cli_lockfile`, `cli_moddetect`, `cli_remove`,
`cli_update`. `cli_pipeline_add.rs` renamed to `cli_add.rs` with the
`--pipeline` argument and the "hidden from help" test removed; new
test asserts each unimplemented command prints its phase tag.
`cli_search` is unchanged. `cli_exit_codes` general-error case
retargeted at `add --global` with HOME unset (same code path,
post-v0.1).

Docs: AGENTS.md module-layout / Install layout / Testing sections
refreshed.

103 unit + 30+ integration tests pass.
